+++ features/fixtures/mazerunner/cxx-scenarios-bugsnag/src/main/java/com/bugsnag/android/mazerunner/scenarios/CXXConfiguredContextScenario.kt
@@ -1,0 +1,22 @@
+package com.bugsnag.android.mazerunner.scenarios
+
+import android.content.Context
+import com.bugsnag.android.Configuration
+
+internal class CXXConfiguredContextScenario(
+    config: Configuration,
+    context: Context,
+    eventMetadata: String?
+) : Scenario(config, context, eventMetadata) {
+    init {
+        System.loadLibrary("cxx-scenarios-bugsnag")
+        config.context = "CustomConfiguredContext"
+    }
+
+    external fun crash()
+
+    override fun startScenario() {
+        super.startScenario()
+        crash()
+    }
+}