+++ CHANGELOG.md
@@ -7,6 +7,9 @@
 * Avoid reporting false-positive background ANRs with improved foreground detection
   [#1429](https://github.com/bugsnag/bugsnag-android/pull/1429)
 
+* Configuration.context is now correctly reported from NDK events
+  [#1437](https://github.com/bugsnag/bugsnag-android/pull/1437)
+
 ## 5.14.0 (2021-09-29)
 
 ### Enhancements